Update Sample App
Rearrange Sample App structure.
